--- src/components/Profile/AboutProfile/AboutProfile.module.css
@@ -1,28 +1,0 @@
-.aboutProfile {
-  background-image: url("https://olympus.crumina.net/wp-content/uploads/buddypress/members/1/cover-image/5c220b12b44cd-bp-cover-image.jpg");
-  height: 300px;
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: cover;
-}
-
-.container {
-  width: 70%;
-  margin: 0 auto;
-}
-
-.aboutProfile__about {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  top: 250px;
-}
-
-.aboutProfile__fullname {
-  background-color: blue;
-  margin-left: 50px;
-  font-size: 24px;
-  font-weight: bold;
-  color: white;
-}
