+++ src/components/Profile/Posts/Posts.module.css
@@ -1,4 +1,4 @@
-.myPosts {
+/* .myPosts {
   max-width: 70%;
   margin: 0 auto;
-}
+} */
