+++ src/components/Profile/Profile.module.css
@@ -1,3 +1,64 @@
-.profile__posts {
-  max-width: 70%;
+.profile {
+  position: relative;
+  background-image: url("https://olympus.crumina.net/wp-content/uploads/buddypress/members/1/cover-image/5c220b12b44cd-bp-cover-image.jpg");
+  height: 300px;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
 }
+
+.profile__overlay {
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.4);
+}
+
+.profile__container {
+  width: 70%;
+  margin: 0 auto;
+}
+
+.profile__about {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  bottom: 20px;
+}
+
+.profile__avatar {
+  border: 3px solid white;
+  width: 150px;
+  height: 150px;
+}
+
+.profile__fullname {
+  font-size: 24px;
+  font-weight: 600;
+  color: white;
+}
+
+
+.container {
+	max-width: 70%;
+	margin: 0 auto;
+}
+
+.wrapper {
+	display: flex;
+  justify-content: space-between;
+}
+
+.profile__info {
+  background-color: rgba(63, 66, 87, 0.1);
+  padding: 20px 0;
+}
+
+.info {
+  flex: 1;
+  margin-right: 20px;
+}
+
