+++ src/components/Profile/ProfileContainer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Profile from './Profile'
+import { Profile } from './Profile'
 import { connect } from 'react-redux'
 import { compose } from 'redux'
 import {
